Fix ordinal references in constructor args
